refactor(mqtt): drop the last schema_0 import from the transport bootstrap
Move log_schema_drift out of _impl/schema_0/field_metadata.py into a new
top-level span_panel_api/schema_drift.py — it only diffs two
HomieSchemaTypes dicts and never touched the schema_0 TYPE_* constants,
so it was already schema-agnostic and belongs at the bootstrap level, not
behind the adapter seam. mqtt/client.py now imports it from there instead
of from _impl/schema_0, so the transport no longer reaches into the
adapter package for anything but its default adapter_factory
(SchemaZeroAdapter), which stays as the plan-sanctioned default.
build_field_metadata and the TYPE_* imports are untouched; the now-unused
_LOGGER/import logging left behind in field_metadata.py are removed.

Also documents two behaviors that were previously implicit or wrong:
- SpanMqttClient.adapter's docstring now notes that transport rebuild
  replaces the adapter instance, so property callbacks registered on the
  old instance do not survive and must be re-registered.
- SchemaAdapter's docstring no longer claims every method is called by
  SpanMqttClient — find_node_by_type and register_property_callback are
  never called from the bootstrap; they exist for external consumers.
